rename shapes directory to models (more generic)
