+++ src/common/registeredShapes.js
@@ -1,18 +1,18 @@
-import Circle from '../shapes/Circle'
-import Epicycloid from '../shapes/Epicycloid'
-import FractalSpirograph from '../shapes/fractal_spirograph/FractalSpirograph'
-import Heart from '../shapes/Heart'
-import Hypocycloid from '../shapes/Hypocycloid'
-import InputText from '../shapes/input_text/InputText'
-import LSystem from '../shapes/lsystem/LSystem'
-import Polygon from '../shapes/Polygon'
-import Reuleaux from '../shapes/Reuleaux'
-import Rose from '../shapes/Rose'
-import SpaceFiller from '../shapes/space_filler/SpaceFiller'
-import Star from '../shapes/Star'
-import TessellationTwist from '../shapes/tessellation_twist/TessellationTwist'
-import V1Engineering from '../shapes/v1_engineering/V1Engineering'
-import Wiper from '../shapes/Wiper'
+import Circle from '../models/Circle'
+import Epicycloid from '../models/Epicycloid'
+import FractalSpirograph from '../models/fractal_spirograph/FractalSpirograph'
+import Heart from '../models/Heart'
+import Hypocycloid from '../models/Hypocycloid'
+import InputText from '../models/input_text/InputText'
+import LSystem from '../models/lsystem/LSystem'
+import Polygon from '../models/Polygon'
+import Reuleaux from '../models/Reuleaux'
+import Rose from '../models/Rose'
+import SpaceFiller from '../models/space_filler/SpaceFiller'
+import Star from '../models/Star'
+import TessellationTwist from '../models/tessellation_twist/TessellationTwist'
+import V1Engineering from '../models/v1_engineering/V1Engineering'
+import Wiper from '../models/Wiper'
 
 /*----------------------------------------------
 Supported input shapes
