+++ src/models/Epicycloid.js
