+++ src/models/Hypocycloid.js
