+++ src/models/Polygon.js
