+++ src/models/Star.js
