+++ src/models/input_text/raysol_cursive.js
