+++ src/models/input_text/raysol_cursive.txt
